+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 5.0.0 (Unreleased)
+* [Classlike, function and constant completions now show their FQSEN in the `detail` property instead of the `label`](https://gitlab.com/Serenata/Serenata/issues/269)
 * Fix namespace autocompletion suggestions not having their backslashes properly escaped
 * Fix classlike autocompletion suggestions not having their backslashes properly escaped when starting with a leading slash
 * See also the release notes for versions [5.0.0-RC](https://gitlab.com/Serenata/Serenata/-/tags/5.0.0-RC), [5.0.0-RC2](https://gitlab.com/Serenata/Serenata/-/tags/5.0.0-RC2) and [5.0.0-RC3](https://gitlab.com/Serenata/Serenata/-/tags/5.0.0-RC3).